project organization
- Top-level content is now in the `general` module.
- Changed `core` to `main`
