+++ website/general/templates/404.j2
@@ -1,4 +1,4 @@
-{% extends "base.j2" %}
+{% extends "layout/base.j2" %}
 
 {% block title %}404 Page Not Found{% endblock %}
 
@@ -6,6 +6,6 @@
 
 <h2 class="main">404 Page Not found</h2>
 <p>Sorry, the page you are looking for does not exist.</p>
-<a href="{{ url_for('home') }}" class="button big">Return to the homepage</a>
+<a href="{{ url_for('main.home') }}" class="button big">Return to the homepage</a>
 
 {% endblock %}
