+++ website/general/templates/components/header.j2
@@ -1,5 +1,5 @@
 <header id="header">
-    <a href="{{ url_for('core.home') }}" class="logo"><strong>Dr. Jessica De
+    <a href="{{ url_for('main.home') }}" class="logo"><strong>Dr. Jessica De
             Silva</strong> California State University, Stanislaus</a>
     <ul class="icons">
         <li><img src="{{ url_for('static', filename='images/stanlogo.png') }}" width="10%"
