+++ website/general/templates/components/sidebar.j2
@@ -1,7 +1,7 @@
 <div id="sidebar">
     <div class="inner">
 
-        <!-- Menu -->
+        {# Menu #}
         {% include "components/nav.j2" %}
 
         <!-- Section -->
@@ -16,7 +16,7 @@
             </ul>
         </section>
 
-        <!-- Footer -->
+        {# Footer #}
         {% include "components/footer.j2" %}
 
     </div>
