+++ website/general/templates/layout/base.j2
@@ -12,44 +12,32 @@
 <html lang="en-US">
 
     <head>
-        <!-- Required meta tags -->
+        {# Required meta tags #}
         <meta charset="UTF-8" />
         <meta name="viewport" content="width=device-width, initial-scale=1, user-scalable=no" />
         <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-
-        <!-- Page title -->
-        <title>
-            {% block title %} {% endblock %}
-        </title>
-
-        <!-- CSS -->
+        {# Page title #}
+        <title>{% block title %}{% endblock %}</title>
+        {# Local CSS file for styling the application #}
         <link rel="stylesheet" href="{{ url_for('static', filename='css/main.css') }}" />
-
-        <!-- Additional Styling -->
-        {% block additional_styling %} {% endblock %}
+        {# Additional styling #}
+        {% block additional_styling %}{% endblock %}
     </head>
 
     <body class="is-preload">
-
         <!-- Wrapper -->
         <div id="wrapper">
-
             <!-- Main -->
             <div id="main">
                 <div class="inner">
-
-                    <!-- Header -->
+                    {# Page header #}
                     {% include "components/header.j2" %}
-
-                    <!-- Main Content -->
+                    {# Page content #}
                     {% block main_content %} {% endblock %}
-
                 </div>
             </div>
-
-            <!-- Sidebar -->
+            {# Sidebar #}
             {% include "components/sidebar.j2" %}
-
         </div>
 
         <!-- Scripts -->
@@ -58,7 +46,8 @@
         <script src="{{ url_for('static', filename='js/breakpoints.min.js') }}"></script>
         <script src="{{ url_for('static', filename='js/util.js') }}"></script>
         <script src="{{ url_for('static', filename='js/main.js') }}"></script>
-
+        {# Additional scripts #}
+        {% block additional_scripts %}{% endblock %}
     </body>
 
 </html>
