+++ website/general/templates/starter_page.j2
@@ -1,11 +1,7 @@
 {% extends "layout/base.j2" %}
 {% block title %}
-
 {# Add page title here #}
-
 {% endblock %}
 {% block main_content %}
-
 {# Add page content here #}
-
 {% endblock %}
